backend/pyproject.toml
[project]
name = "shopdjango"
version = "0.1.0"
description = "My final project"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "django>=5.2.3",
    "django-filter>=25.1",
    "django-guardian>=3.0.0",
    "djangorestframework>=3.16.0",
    "markdown>=3.8.2",
    "pygments>=2.19.1",
    "pytest-django>=4.11.1",
    "pyyaml>=6.0.2",
    "uritemplate>=4.2.0",
    "psycopg[binary]>=3.1",
    "djangorestframework-simplejwt>=5.5.0",
    "django-allauth[socialaccount]>=65.9",
    "drf-spectacular>=0.28",
    "django-extensions>=4.1",
    "django-cors-headers>=4.7.0",
    "pillow>=11.2.1",
    "stripe>=12.3.0",
    "python-dotenv>=1.1.1",
    "celery>=5.5.3",
    "redis>=6.4.0",
    "jinja2>=3.1.6",
    "pdfkit>=1.0.0",
    "flask>=3.1.2",
    "django-storages>=1.14.6",
    "boto3>=1.40.24",
]

[dependency-groups]
dev = [
    "ruff>=0.12.1",
]
